Move scroll delegate methods to own extension
